fix:/restrucre clock component
