+++ src/clock/clock.css
@@ -1,0 +1,205 @@
+@import url('https://fonts.googleapis.com/css2?family=Open+Sans:wght@400;600;700&display=swap');
+
+.wrapper{
+    box-sizing: border-box;
+  
+    position: fixed;
+    top: 0px;
+    left: 0px;
+    display: flex;
+    flex-direction: row;
+    
+    height: 100%;
+    width: calc(100% - 71px);
+  }
+  
+  
+  
+  .settings-pane{
+    left: 0;
+    box-sizing: border-box;
+    display: inline-flex;
+    height: auto;
+    padding: 40px;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 40px;
+  
+    border-radius: 0px 20px 20px 0px;
+    background: var(--White, #FFF);
+    box-shadow: 5px 5px 10px 0px rgba(71, 88, 220, 0.15);
+  
+  }
+  
+  /*Add animation*/
+  
+  /*Horizontal Slide*/
+  @keyframes slideInX {
+    from {
+      transform: translateX(-100%);
+    }
+    to {
+      transform: translateX(0);
+    }
+  }
+  
+  /*Vertical Slide*/
+  @keyframes slideInY {
+    from {
+      transform: translateY(-100%);
+    }
+    to {
+      transform: translateY(0);
+    }
+  }
+  
+  .settings-pane.slide-in{
+    animation: slideInX 0.5s ease-in-out;
+  }
+  
+  
+  
+  .settings-pane-title{
+    color: var(--RoyalBlue, #4758DC);
+    font-family: "Open Sans";
+    font-size: 18px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 28px; /* 155.556% */
+    
+  }
+  
+  .settings{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 20px;
+  }
+  
+  
+  .focus-duration, .break-duration, .total-session-count{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: flex-start;
+    gap: 4px;
+  }
+  
+  .input-label{
+    color: var(--RoyalBlue, #4758DC);
+    font-family: "Open Sans";
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: normal;
+  }
+  
+  input{
+    box-sizing: border-box;
+    width: 110px;
+    height: 28px;
+    padding: 8px;
+    border-radius: 4px;
+    border: 1px solid var(--Manatee, #97999C);
+  }
+  
+  input:hover{
+    border-color: #4758DC;
+  }
+  
+  input:active{
+    border: 2px solid var(--RoyalBlue, #4758DC);
+  }
+  
+  
+  .timer{
+    box-sizing: border-box;
+  
+    position: relative;
+    flex-grow: 1;
+    height: 100vh;
+    width: 100vw;
+    
+    display: flex;
+    padding: 20px;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 40px;
+  
+    align-self: stretch;
+  }
+  
+  
+  .remaining-time{
+    color: var(--Black, #000);
+    font-family: "Open Sans";
+    font-size: 72px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: normal;
+  }
+  
+  .session-display{
+    color: var(--RoyalBlue, #4758DC);
+    font-family: "Open Sans";
+    font-size: 24px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: normal;
+  }
+  
+  .timer-buttons{
+    display: flex;
+    align-items: center;
+    gap: 16px;
+  }
+  
+  
+  @media only screen and (max-width: 767px) {
+  
+    .Wrapper{
+      position: fixed;
+      top: 61px;
+      left: 0px;
+  
+      flex-direction: column;
+  
+      width: 100%;
+      height: calc(100% - 61px);
+    }
+    
+    
+    .settings-pane{
+      display: flex;
+      box-sizing: border-box;
+      height: auto;
+      width:100vw;
+      padding: 40px 0px 60px  0px;
+    
+      border-radius: 0px 0px 20px 20px;
+      box-shadow: 0px 10px 10px 0px rgba(71, 88, 220, 0.15);
+    }
+  
+    .settings-pane.slide-in{
+      animation: slideInY 0.5s ease-in-out;
+    }
+    
+    .settings-pane-title{
+      font-size: 14px;
+    }
+    
+    .input-label{
+      font-size: 12px;
+    }
+  
+    .remaining-time{
+      font-size: 40px;
+    }
+  
+    .session-display{
+      font-size: 16px;
+    }
+   
+  }
