+++ src/navbar/navbar.css
@@ -4,37 +4,24 @@
     position: fixed;
     top: 0;
     left: 0;
-    height: 100vh;
-
-    width: auto;
-    
+    height: auto;
+    width: 100vw;
 
     display: flex;
-    padding: 60px 30px;
-    flex-direction: column;
+    padding: 10px 20px;
+    flex-direction: row;
     justify-content: space-between;
     align-items: center;
     align-self: stretch;
     
     background: var(--White, #FFF);
-    border-right: 2px solid var(--RoyalBlue, #4758DC);
+    border-bottom: 2px solid var(--RoyalBlue, #4758DC);
 
     z-index:1000;
 }
-
-@media only screen and (max-width: 767px) {
-    
-    .navbar {
-
-        height: auto;
-        width: 100%;
-        padding: 10px 20px;
-
-        flex-direction: row;
-
-        border-right: none;
-        border-bottom: 1px solid var(--RoyalBlue, #4758DC);
-    }
-
-   
-  }
+.actions{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 10px;
+}
