+++ src/textButton/textButton.jsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import Button from '@mui/material/Button';
 import { colors } from '../colors/colors';
 
-export default function TextButton({label}) {
+export default function TextButton({label, onClick}) {
     return (
         <Button style={{
             color: colors.RoyalBlue,
